+++ pyproject.toml
@@ -2,9 +2,7 @@
 name = "grandma-gcn"
 version = "0.1.0"
 description = "Default template for PDM package"
-authors = [
-    {name = "Roman", email = "roman.lemontagner@gmail.com"},
-]
+authors = [{ name = "Roman", email = "roman.lemontagner@gmail.com" }]
 dependencies = [
     "gcn-kafka>=0.3.3",
     "tomli>=2.0.2",
@@ -16,7 +14,7 @@ dependencies = [
     "astropy-healpix>=1.1.2",
     "celery[redis]>=5.5.2",
     "dotenv>=0.9.9",
-    "gwemopt @ git+https://github.com/FusRoman/old_gwemopt@2daa8931271af4feca65bf827742acefc19df3c7",
+    "gwemopt @ git+https://github.com/FusRoman/old_gwemopt@8632efe7baf9212e9d490dc0bb4fccee388dc83e",
 ]
 requires-python = ">=3.12"
 readme = "README.md"
@@ -31,10 +29,7 @@ build-backend = "pdm.backend"
 distribution = true
 
 [project.optional-dependencies]
-test = [
-    "pytest>=8.3.3",
-    "pytest-mock>=3.14.0",
-]
+test = ["pytest>=8.3.3", "pytest-mock>=3.14.0"]
 
 lint = ["black>=24.10.0", "pre-commit>=4.0.1", "pylint>=3.3.2"]
