+++ src/grandma_gcn/worker/gwemopt_worker.py
@@ -5,7 +5,7 @@ from grandma_gcn.gcn_stream.gw_alert import GW_alert
 from grandma_gcn.worker.celery_app import celery
 import logging
 
-from grandma_gcn.worker.gwemopt_launcher import init_gwemopt
+from grandma_gcn.worker.gwemopt_init import init_gwemopt
 
 
 def setup_task_logger(task_name: str) -> logging.Logger:
@@ -58,7 +58,7 @@ def run_gwemopt(
     Run the gwemopt observation plan.
 
     Parameters
-    ----------100
+    ----------
     gw_alert : GW_alert
         The GW_alert object containing the alert information.
     telescopes : list[str]
@@ -71,6 +71,12 @@ def run_gwemopt(
         The observation strategy to use.
     logger : logging.Logger
         Logger to use for logging. If None, a new logger will be created.
+
+    Returns
+    -------
+    tuple[Table, Any]
+        - tiles_tables: the table of tiles generated by gwemopt
+        - galaxies_table: the table of galaxies generated by gwemopt
     """
     logger.info("Flattening the skymap...")
     flat_map = gw_alert.flatten_skymap(nside)
@@ -92,7 +98,7 @@ def run_gwemopt(
     )
 
     logger.info("gwemopt initialized. Running observation plan...")
-    _ = gw_alert.run_observation_plan(
+    obs_plan_results = gw_alert.run_observation_plan(
         telescopes,
         params,
         map_struct,
@@ -100,6 +106,8 @@ def run_gwemopt(
         observation_strategy,
     )
     logger.info("Observation plan completed.")
+
+    return obs_plan_results
 
 
 @celery.task(name="gwemopt_task")
@@ -146,7 +154,7 @@ def gwemopt_task(
         logger = setup_task_logger("gwemopt_task_{}".format(gw_alert.event_id))
         logger.info("Starting gwemopt_task...")
 
-        run_gwemopt(
+        _, _ = run_gwemopt(
             gw_alert,
             telescopes,
             nb_tiles,
